+++ CHANGELOG.md
@@ -7,11 +7,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- Empty
+### Added
 
-## [0.2.0] - 2023-01-28
+- Examples updated.
 
-## Added
+## [v0.3.0] - 2023-01-30
+
+### Added
+
+- Shadowing: methods on `State`s that are always executed on a `Stack`'s `tick`,
+even if the `State` is not at the top.
+  - `on_shadow_tick` for when the stack gets ticked through its `tick` method.
+- Added a test for the new shadow feature.
+- Minor fixes in documentation.
+
+## [v0.2.0] - 2023-01-28
+
+### Added
 
 - `Trans`
   - `Replace`: pops and pushes a new state, effectively performing a 
@@ -20,11 +32,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   only one on the stack.
   - Macros and functions on `Stack` for the new transitions.
 
-## [0.1.0] - 2023-01-27
+## [v0.1.0] - 2023-01-27
 
 - First public release of `solstack`.
 
-## Added
+### Added
 
 - Fully capable state stack machine.
 - `State` callbacks added on this version: